+++ CHANGELOG.md
@@ -10,6 +10,7 @@ TODO: Date
   * Removed "styles" and "topiary" features. CSS scoping will be enabled through more general means.
   * Reworked generated component interface
   * Upgraded `lignin` and `lignin-schema` dependencies to 0.0.2 each
+
 * Features:
   * You can now prefix constructor arguments with an explicit visibility (`priv`, `pub`, `pub(restriction)`) to capture them as component instance fields.
   * `bump` resolution is now more reliable in cases where the macro input is constructed in multiple macro contexts.
@@ -34,6 +35,9 @@ TODO: Date
 
   * Optional arguments: `pattern?: Type`
   * Default parameters: `pattern: Type = default`
+  * Conditional attributes: `."attribute-name"? = {Option<&'bump str>}`
+  * Conditional parameters (like conditional attributes)
+  * Boolean attributes: `."attribute-name"? = {bool}`
 
 * Revisions:
   * Improved `Counter` example in the README.